+++ team_fridge/public/css/setting/setting_thema.css
@@ -1,4 +1,4 @@
-/* 컬러서클 */
+/* 라이트 모드 */
 :root[color-theme="light"] {
   --background1: #e7e5e5;
   --background2: #f4f4f4;
@@ -7,7 +7,7 @@
   --boxColor2: #4e4e4e;
   --boxColor3: #858585;
 }
-
+/* 다크모드 */
 :root[color-theme="dark"] {
   --background1: #363535;
   --background2: #474646;
@@ -16,7 +16,9 @@
   --boxColor2: #a09e9e;
   --boxColor3: #7e7b7b;
 }
+
 /* 설정창 */
+body,
 div.main {
   background: var(--background1);
 }
@@ -31,12 +33,14 @@ div.body li {
 }
 
 section,
-div.control {
+nav.control {
   background: var(--background2);
 }
 
-div.main header i,
-div.control li {
+/* 아이콘 색 */
+i.fa.fa-cog,
+i.fa.fa-bell,
+i.fa.fa-calendar.check {
   color: var(--boxColor2);
 }
 
@@ -73,19 +77,20 @@ div li.add {
   background: var(--background2);
 }
 
-/* 로그인 창 */
+/* 로그인 & 회원가입 창 */
+.join_body h2,
 .login_body h2 {
   color: var(--boxColor);
 }
-
+article.join_body button,
 article.login_body button {
   background: var(--background3);
 }
-
+article.join_body input,
 article.login_body input {
   border: 2px solid var(--boxColor3);
 }
-
+article.join_body input:focus,
 article.login_body input:focus {
   border: 2px solid var(--boxColor2);
 }
