+++ team_fridge/public/js/setting/setting_join.js
