+++ team_fridge/public/js/setting/setting_menu.js
