+++ team_fridge/views/setting/login.pug
@@ -1,9 +1,11 @@
 extends ../layout_home
 block content
-    link(rel="stylesheet", href="/css/setting.css")
-    link(rel="stylesheet", href="/css/setting_login.css")
-    script(src="/js/setting_menu.js")
-    script(src="/js/setting_login.js")
+    link(rel="stylesheet", href="/css/setting/setting.css")
+    link(rel="stylesheet", href="/css/setting/setting_login.css")
+    link(rel="stylesheet", href="/css/setting/setting_thema.css")
+    script(src="/js/setting/setting_menu.js")
+    script(src="/js/setting/setting_login.js")
+    script(src="/js/setting/setting_thema.js")
     div.main
       div.body
         section.main
@@ -12,7 +14,7 @@ block content
             form.login(method="post")
               if NEED
                   div.error= NEED
-              input.userid(name="ps_id" placeholder="아이디")
-              input.password(id="password" name="ps_pw" placeholder="비밀번호" type="password")
+              input.userid(name="u_id" placeholder="아이디")
+              input.password(id="password" name="u_pw" placeholder="비밀번호" type="password")
               button.btn_login(type="button") 로그인
               button.btn_join(type="button") 회원가입
