+++ team_fridge/views/setting/setting.pug
@@ -1,11 +1,11 @@
 extends ../layout_home
 
 block content 
-    link(rel="stylesheet", href="/css/setting.css")
-    link(rel="stylesheet", href="/css/setting_alam.css")
-    link(rel="stylesheet", href="/css/setting_thema.css")
-    script(src="/js/setting_menu.js")
-    script(src="/js/setting_thema.js")
+    link(rel="stylesheet", href="/css/setting/setting.css")
+    link(rel="stylesheet", href="/css/setting/setting_alam.css")
+    link(rel="stylesheet", href="/css/setting/setting_thema.css")
+    script(src="/js/setting/setting_menu.js")
+    script(src="/js/setting/setting_thema.js")
     div.main
         div.body
             section.box
